Refine affine_param_gen logic
